Refactor: Split main.go
