+++ api/src/main.go
@@ -2,11 +2,8 @@ package main
 
 import (
 	"encoding/json"
-	"gorm.io/gorm"
 	"errors"
 	"fmt"
-	"time"
-	"regexp"
 	"io"
 	"log"
 	"net/http"
@@ -14,8 +11,6 @@ import (
 	"os"
 	"github.com/gin-gonic/gin"
 	"github.com/gin-contrib/cors"
-	"strconv"
-	"github.com/jinzhu/now"
 )
 
 type Token struct {
@@ -30,34 +25,12 @@ type RequestData struct {
 	Uid  string `json:"uid"`
 }
 
-type UserRequestData struct {
-	Uid string `json:"uid"`
-	Login string `json:"login"`
-	Wallet string `json:"wallet"`
-}
-
-type ActivityRequestData struct {
-	Mac string `json:"mac"`
-	Uid string `json:"uid"`
-}
-
 type RoleRequestData struct {
 	Name string `json:"name"`
 }
 
 type LocationRequestData struct {
 	Name string `json:"name"`
-}
-
-type M5StickRequestData struct {
-	Mac string `json:"mac"`
-	RoleName string `json:"role"`
-	LocationName string `json:"location"`
-}
-
-type Schedule struct {
-	Date string `json:"date"`
-	Login []string `json:"login"`
 }
 
 type UserData struct {
@@ -94,7 +67,7 @@ func main() {
 	router.POST("/shift", addShiftData)
 
 	router.POST("/activities", addActivity)
-	router.GET("/activities/cleanings", getCleanData)
+	router.GET("/activities/cleanings", getActivityCleanData)
 
 	router.POST("/roles", addRole)
 
@@ -106,106 +79,6 @@ func main() {
 	router.PUT("/users", editUser)
 
 	router.Run(":" + os.Getenv("PORT"))
-}
-
-func getShiftData(c *gin.Context) {
-	date, err := getQueryAboutDate(c)
-	if err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid query"})
-		return
-	}
-
-	//roleがcleaningのactivityを取得
-	shifts, err := getShiftFromDB(date)
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to get shift"})
-		return
-	}
-	c.JSON(http.StatusOK, shifts)
-}
-
-func addShiftData(c *gin.Context) {
-	var schedule []Schedule
-
-	if err := c.BindJSON(&schedule); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-	if len(schedule) == 0 {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Shift is required"})
-		return
-	}
-	for _, s := range schedule {
-		log.Println(s.Date)
-		for _, l := range s.Login {
-			log.Println(l)
-		}
-	}
-	if date, err := addShiftToDB(schedule); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	} else {
-		c.JSON(http.StatusOK, gin.H{"date": date})
-	}
-}
-
-func getCleanData(c *gin.Context) {
-	//start_timeとend_timeを取得
-	start_time, end_time, err := getQueryAboutTime(c)
-	if err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid query"})
-		return
-	}
-
-	//roleがcleaningのactivityを取得
-	Activities, err := getActivitiesFromDB(start_time, end_time, "cleaning")
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to get activities"})
-		return
-	}
-	c.JSON(http.StatusOK, Activities)
-}
-
-func getQueryAboutDate(c *gin.Context) (string, error) {
-	date := c.Query("date")
-	if date == "" {
-		date = time.Now().Format("2006-01-02")
-	}
-	if !regexp.MustCompile(`^\d{4}-\d{2}-\d{2}$`).MatchString(date) {
-		return "", errors.New("Invalid date format. It should be in YYYY/MM/DD format")
-	}
-	return date, nil
-}
-
-func getQueryAboutTime(c *gin.Context) (int64, int64, error) {
-	var start_time int64
-	var end_time int64
-	var err error
-
-	start := c.Query("start")
-	if start == "" {
-		//startパラメータがない場合は当日の0時0分0秒を取得
-		start_time = now.BeginningOfDay().Unix()
-	} else {
-		start_time, err = strconv.ParseInt(start, 10, 64)
-		if err != nil {
-			return 0, 0, err
-		}
-	}
-	end := c.Query("end")
-	if end == "" {
-		//endパラメータがない場合はstart_timeの24時間後を取得
-		end_time = start_time + 24*60*60
-	} else {
-		end_time, err = strconv.ParseInt(end, 10, 64)
-		if err != nil {
-			return 0, 0, err
-		}
-	}
-	if start_time >= end_time {
-		return 0, 0, errors.New("Invalid time range")
-	}
-	return start_time, end_time, nil
 }
 
 // 環境変数の読み込み
@@ -361,60 +234,6 @@ func fetchUserData(accessToken string) (*UserData, error) {
 	return &UserData{IntraName: intraName}, nil
 }
 
-func addActivity(c *gin.Context) {
-	var requestData ActivityRequestData
-	
-	// JSONリクエストボディを解析してrequestDataに格納
-	if err := c.BindJSON(&requestData); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-	// requestDataが空でないことを確認（MacとUidが非空の文字列）
-	if requestData.Mac == "" || requestData.Uid == "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "All parameters are required"})
-		return
-	}
-	db, err := connectToDB()
-	if err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		log.Fatal("Failed to initialize database:", err)
-		return
-	}
-	user := User{}
-	if err := db.Where("uid = ?", requestData.Uid).First(&user).Error; err != nil {
-		if err == gorm.ErrRecordNotFound {
-			log.Println("User not found:", err)
-			c.JSON(http.StatusNotFound, gin.H{"error": "User not found"})
-		} else {
-			log.Println("Failed to get user:", err)
-			c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to get user"})
-		}
-		return
-	}
-	m5Stick := M5Stick{}
-	if err := db.Where("mac = ?", requestData.Mac).First(&m5Stick).Error; err != nil {
-		if err == gorm.ErrRecordNotFound {
-			log.Println("M5Stick not found:", err)
-			c.JSON(http.StatusNotFound, gin.H{"error": "M5Stick not found"})
-		} else {
-			log.Println("Failed to get M5Stick:", err)
-			c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to get M5Stick"})
-		}
-		return
-	}
-	// Add a new activity
-	activity := Activity{UserID: user.ID, M5StickID: m5Stick.ID}
-	if result := db.Create(&activity); result.Error != nil {
-		log.Fatal("Failed to create activity:", result.Error)
-		c.JSON(http.StatusBadRequest, gin.H{"error": result.Error.Error()})
-		return
-	}
-	// 取得したuserDataを含めてレスポンスを返す
-	c.JSON(http.StatusOK, gin.H{
-		"uid": requestData.Uid, "mac": requestData.Mac})
-	return
-}
-
 func addRole(c *gin.Context) {
 	var requestData RoleRequestData
 
@@ -456,95 +275,5 @@ func addLocation(c *gin.Context) {
 		return
 	}
 	c.JSON(http.StatusOK, gin.H{"name": requestData.Name})
-	return
-}
-
-func addM5Stick(c *gin.Context) {
-	var requestData M5StickRequestData
-
-	// JSONリクエストボディを解析してrequestDataに格納
-	if err := c.BindJSON(&requestData); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-	if requestData.Mac == "" || requestData.RoleName == "" || requestData.LocationName == "" {
-		// すべてのパラメータが必須
-		c.JSON(http.StatusBadRequest, gin.H{"error": "All parameters are required"})
-		return
-	}
-	// データベースにM5Stickを追加
-	if err := addM5StickToDB(requestData.Mac, requestData.RoleName, requestData.LocationName); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-	c.JSON(http.StatusOK, gin.H{"mac": requestData.Mac, "role": requestData.RoleName, "location": requestData.LocationName})
-	return
-}
-
-func addUser(c *gin.Context) {
-	var requestData UserRequestData
-
-	// JSONリクエストボディを解析してrequestDataに格納
-	if err := c.BindJSON(&requestData); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-	if requestData.Login == "" {
-		// Loginは必須
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Login is required"})
-		return
-	}
-	if userExists(requestData.Login) {
-		c.JSON(http.StatusConflict, gin.H{"error": "User with this login already exists"})
-		return
-	}
-	// データベースにUserを追加
-	if err := addUserToDB(requestData.Uid, requestData.Login, requestData.Wallet); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-	response := make(gin.H)
-
-	response["login"] = requestData.Login
-	if requestData.Uid != "" {
-		response["uid"] = requestData.Uid
-	}
-	if requestData.Wallet != "" {
-		response["wallet"] = requestData.Wallet
-	}
-
-	c.JSON(http.StatusOK, response)
-	return
-}
-
-func editUser(c *gin.Context) {
-	var requestData UserRequestData
-
-	// JSONリクエストボディを解析してrequestDataに格納
-	if err := c.BindJSON(&requestData); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-	if requestData.Login == "" {
-		// Loginは必須
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Login is required"})
-		return
-	}
-	// DB上のUserを編集
-	if err := editUserInDB(requestData.Uid, requestData.Login, requestData.Wallet); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-	response := make(gin.H)
-
-	response["login"] = requestData.Login
-	if requestData.Uid != "" {
-		response["uid"] = requestData.Uid
-	}
-	if requestData.Wallet != "" {
-		response["wallet"] = requestData.Wallet
-	}
-
-	c.JSON(http.StatusOK, response)
 	return
 }
